requirements.txt
pywinauto>=0.6.8
Pillow>=8.0.0
pywin32>=301
comtypes>=1.1.0 